+++ CHANGELOG.md
@@ -14,6 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Do not add additional links when editing messages (this is maybe a bug in RocketChat)
 
+## [0.2.0] - 2020-03-18
+
+### Added
+
+* Search for issues on the server
+* i18n Labels for settings
+* Setting to deactivate attachments (default is activated)
+
 ## [0.1.0] - 2020-03-17
 
 ### Added